refactor: rename package historic to history
the package interacts with history it is not itself a thing from the past
