+++ pkg/history/historic_test.go
@@ -1,4 +1,4 @@
-package historic
+package history
 
 import (
 	"testing"
